pyproject.toml
[tool.poetry]
name = "licenses"
version = "0.1.0"
description = ""
authors = ["colrosezx <kirillkovalev-2005@mail.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.7"
pydantic = "^2.10.6"
sqlalchemy = "^2.0.37"
uvicorn = "^0.34.0"
pydantic-settings = "^2.7.1"
alembic = "^1.14.1"
python-dateutil = "^2.9.0.post0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
